@@ -1,18 +1,17 @@
-package com.melkonian.example.lesson3.adapter;
+package com.melkonian.example.lessons.adapter;
 
-import android.support.annotation.NonNull;
-import android.support.v7.widget.CardView;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
-import com.melkonian.example.lesson3.R;
+import androidx.annotation.NonNull;
+import androidx.cardview.widget.CardView;
+import androidx.recyclerview.widget.RecyclerView;
+import com.melkonian.example.lessons.R;
 
 public class CitiesAdapter extends RecyclerView.Adapter<CitiesAdapter.ViewHolder> {
   private String[] data;
 
-  // Для recyclerView
   private OnRecyclerViewClickListener recyclerItemClickListener;
   //private View.OnClickListener clickListener;
 
@@ -40,11 +39,9 @@ public class CitiesAdapter extends RecyclerView.Adapter<CitiesAdapter.ViewHolder
     return data.length;
   }
 
-  public void SetOnItemClickListener(OnRecyclerViewClickListener itemClickListener){
+  public void SetOnItemClickListener(OnRecyclerViewClickListener itemClickListener) {
     this.recyclerItemClickListener = itemClickListener;
   }
-
-
 
   public static class ViewHolder extends RecyclerView.ViewHolder {
     //private TextView textView;
@@ -52,40 +49,14 @@ public class CitiesAdapter extends RecyclerView.Adapter<CitiesAdapter.ViewHolder
     //private View.OnClickListener clickListener;
     private OnRecyclerViewClickListener recyclerItemClickListener;
 
-    //public ViewHolder(TextView view, View.OnClickListener itemClickListener) {
-  /*  public ViewHolder(TextView view, OnRecyclerViewClickListener itemClickListener) {
-      super(view);
-      //this.textView = view;
-      //this.clickListener = itemClickListener;
-      this.recyclerItemClickListener = itemClickListener;
-
-      textView.setOnClickListener(new View.OnClickListener() {
-        @Override public void onClick(View v) {
-          if (recyclerItemClickListener != null) {
-            recyclerItemClickListener.onItemClick(v, getAdapterPosition() + 1);
-          }
-        }
-      });
-
-      //textView.setOnClickListener(itemClickListener);
-      //textView.setOnClickListener(new View.OnClickListener() {
-      //  @Override public void onClick(View v) {
-      //    if (clickListener != null) {
-      //      clickListener.onClick(v);
-      //    }
-      //  }
-      //});
-    }*/
-
     private TextView textView;
+
     public ViewHolder(View view, OnRecyclerViewClickListener itemClickListener) {
       super(view);
       textView = view.findViewById(R.id.text_view);
-      textView.setOnClickListener(new View.OnClickListener() {
-        @Override public void onClick(View v) {
-          if (itemClickListener != null) {
-            itemClickListener.onItemClick(v, getAdapterPosition());
-          }
+      textView.setOnClickListener(v -> {
+        if (itemClickListener != null) {
+          itemClickListener.onItemClick(v, getAdapterPosition());
         }
       });
     }
